tests/wpt/tests/html/cross-origin-embedder-policy/non-initial-about-blank.https.html

<!doctype html>
<meta charset=utf-8>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(t => {
  const frame = document.createElement("iframe");
  t.add_cleanup(() => frame.remove());
  let i = 0;
  frame.onload = t.step_func(() => {
    i++;
    assert_equals(frame.contentDocument.URL, "about:blank");
    frame.src = "about:blank";
    if (i == 2) {
      t.done();
    }
  });
  document.body.append(frame);
}, "Cross-Origin-Embedder-Policy and about:blank");
</script>
